src/requirement_fastAPI.txt
python-dotenv
requests
SQLAlchemy
pandas
psycopg2-binary
fastapi
pydantic
uvicorn